Implement the low-memory rendering pipeline.
Includes some fixes to the various stages.

The low-memory rendering pipeline is enabled by default. This does not
change the decoding output, and has minor effects on speed (that can
likely be compensated for by better optimizing the various stages and
removing remaining non-rp code).

Two TODOs are still left before being able to fully remove the non-rp
code: DC upsampling and encoder-side roundtrip still use code that
belongs to the non-rendering-pipeline path.

```
x86 baseline: 54.481 MP/s
ARM baseline: geomean: 9.14383 MP/s [9.08262, 9.23499]

x86 after: 53.598 MP/s
ARM after: geomean: 9.25677 MP/s [9.21585, 9.31553]
```

+++ lib/jxl/render_pipeline/stage_xyb.cc
@@ -131,7 +131,7 @@ class XYBStage : public RenderPipelineStage {
     msan::UnpoisonMemory(row0 + xsize, sizeof(float) * (xsize_v - xsize));
     msan::UnpoisonMemory(row1 + xsize, sizeof(float) * (xsize_v - xsize));
     msan::UnpoisonMemory(row2 + xsize, sizeof(float) * (xsize_v - xsize));
-    for (int64_t x = -xextra; x < (int64_t)(xsize + xextra); x += Lanes(d)) {
+    for (ssize_t x = -xextra; x < (ssize_t)(xsize + xextra); x += Lanes(d)) {
       const auto in_opsin_x = Load(d, row0 + x);
       const auto in_opsin_y = Load(d, row1 + x);
       const auto in_opsin_b = Load(d, row2 + x);
